002. run and connect to a Postgres instance/requirements.txt
psycopg2-binary==2.9.8